requirements.txt
streamlit
opencv-python-headless
ultralytics
Pillow
numpy
streamlit_webrtc
torchvision
av